+++ src/com/santrong/tcp/client/LocalTcp31008.java
@@ -12,12 +12,12 @@ import com.santrong.util.XmlReader;
  */
 public class LocalTcp31008 extends LocalTcpBase {
 	
-	private int freeSize;// 设置磁盘空间阀值，磁盘空间少于此值则自行停止录制，单位M
+	private long freeSize;// 设置磁盘空间阀值，磁盘空间少于此值则自行停止录制，单位M
 	private int maxTime;// 最大录制时长，单位：分钟
 	
 	private int resultCode;
 
-	public void setFreeSize(int freeSize) {
+	public void setFreeSize(long freeSize) {
 		this.freeSize = freeSize;
 	}
